Refactor callings of method 'render' to proper 'draw' method
Instead of calling the 'render' from the start method, now it is
being called everytime the 'draw' function is called. This is
the proper way to do it, following the documentation of the
method 'draw':

'Refresh the screen and redraw the menu. Should be called whenever
something changes that needs to be redrawn.'

That's exactly the behaviour we expect for a 'render' function